created separate directories for client and server
